XelloxShop-main/requirements.txt
Flask
grok
sentence-transformers
faiss-cpu
numpy
torch
scikit-learn
openai
requests
